feat: support int16/int32/int64 integers

+++ crates/compiler/src/core.rs
@@ -28,7 +28,7 @@ pub enum Expr {
         ty: Ty,
     },
     EInt {
-        value: i32,
+        value: i64,
         ty: Ty,
     },
     EString {
